@@ -1,6 +1,6 @@
 package ru.develgame.jsfgame.chat.entity;
 
-import javax.persistence.*;
+import jakarta.persistence.*;
 import java.io.Serializable;
 import java.util.Date;
 
