@@ -1,7 +1,7 @@
 package ru.develgame.jsfgame.common.jsf;
 
-import javax.enterprise.context.ApplicationScoped;
-import javax.inject.Named;
+import jakarta.enterprise.context.ApplicationScoped;
+import jakarta.inject.Named;
 
 @Named("menu")
 @ApplicationScoped
